requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
astunparse==1.6.3
asyncio-dgram==2.2.0
attrs==24.2.0
braceexpand==0.1.7
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
croniter==3.0.3
discord-ext-menus==1.1
discord-py-interactions==5.13.1
discord-typings==0.9.0
discord.py==2.4.0
dnspython==2.6.1
emoji==2.12.1
mcstatus
frozenlist==1.4.1
humanfriendly==10.0
idna==3.8
import_expression==2.1.0.post1
jishaku @ git+https://github.com/Gorialis/jishaku@0c16ddca174726966704a8f12f6e8f0e26528202
mcstatus==11.1.1
multidict==6.0.5
pillow==10.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
setuptools==72.1.0
six==1.16.0
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
wheel==0.43.0
yarl==1.9.4
